Add files via upload
Problem Statement
The primary objective of the problem for which we sought a solution in this study is to use videos or images to determine whether a fire has occurred in a particular location. In comparison to much more established sensor-based flame recognition techniques, camera-based flame detection techniques have acquired increasing footing. The need for video processing in private, commercial, and forest settings has precipitously increased the use of vision-based fire identification systems. Several fire-related disasters have occurred in recent years as a result of inadequate monitoring or failure to cover vulnerable areas, such as prohibited forest areas or industrial operations.

Dataset
The dataset contains a total of 5000 training images, of which 2500 are fire-affected areas and 2500 are non-fire-affected areas, indicating that the dataset is evenly distributed. In order to improve the precision of our model, we can, if necessary, manually select images from the internet. The following are the data set's specifications:

● Link- https://www.kaggle.com/datasets/mohnishsaiprasad/forest-fire-images

Goal / Scope
Using this technique, we will attempt to detect if a forest region is experiencing a fire outbreak in order to limit the spread of forest fires through early detection. Our study would involve the development of a novel CNN-based model that would be trained on the aforementioned dataset in order to detect the onset of a forest fire at an early stage so that its further spread can be halted.

Expected Results
Idealistically, the deep learning algorithm we develop would be able to classify images into the following categories based on the presence or absence of fire-related characteristics. Typical program would include:

Fire - Forest Fire Outbreak Detected
Neutral - No Forest Fire Outbreak Has Been Identified